pyproject.toml
[project]
name = "cafe-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "django-filter (>=24.3,<25.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
